requirements.txt
streamlit 
pandas
matplotlib
plotly
numpy
sklearn.linear_model
sklearn.preprocessing
base64